# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Dot Net Lab REST APIs",
    "description": "This is the APIs documentation for Dot Net Lab Web App.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Group ??",
      "email": "benlun99999@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:40000/book-shop",
      "description": "This is the APIs documentation for Dot Net Lab Web App."
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "Books APIs"
        ],
        "description": "Get all books",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/carts/{userId}": {
      "get": {
        "tags": [
          "Carts APIs"
        ],
        "description": "Get user's carts",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/carts": {
      "post": {
        "tags": [
          "Carts APIs"
        ],
        "description": "Create cart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/carts/{cartId}": {
      "post": {
        "tags": [
          "Carts APIs"
        ],
        "description": "Add cart item",
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/carts/{cartId}/items/{bookId}": {
      "delete": {
        "tags": [
          "Carts APIs"
        ],
        "description": "Remove cart item",
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users APIs"
        ],
        "description": "Get all users",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CartItemRequest": {
        "required": [
          "bookId"
        ],
        "type": "object",
        "properties": {
          "bookId": {
            "minLength": 1,
            "type": "string"
          },
          "quantity": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateCartRequest": {
        "required": [
          "items",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "minLength": 1,
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemRequest"
            }
          }
        },
        "additionalProperties": false
      }
    }
  }
}